requirements.txt
fastapi
uvicorn
groq
python-dotenv
pypdf
sentence-transformers
faiss-cpu
streamlit
requests
